wine_data_enricher/requirements.txt
pandas==2.2.0
openpyxl==3.1.0
openai>=2.0.0
duckduckgo-search>=8.1.0
python-dotenv==1.0.0
